backend/requirements.txt
Django>=4.2.17,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.0,<1.1
amadeus>=9.0.0
django-cors-headers>=4.3.1,<4.4
Pillow==9.5.0
requests>=2.31.0,<2.32
openai>=1.0.0
pytest>=7.3.1,<7.4
pytest-django>=4.5.2,<4.6
locust>=2.37.10


